Follow up to use Core Data
In this revision the JSON data is only download once if there are no Core Data entities. Once the JSON has been downloaded and decoded into the previous structs we iterate over them and create Core Data entities the will be persisted. The structs from before are discarded after creating the Core Data entities. The modifications to the app were fairly minor. The main view was modified to use the FilteredList view from earlier and the data types were changed to use the Core Data entities and their wrapped computed properties. Structurally, it stayed pretty much the same.